Conflicts:
	src/beingmanager.cpp
	src/gui/confirm_dialog.cpp
	src/gui/inventorywindow.cpp
	src/gui/inventorywindow.h
	src/gui/label.cpp
	src/gui/label.h
	src/gui/popup.cpp
	src/gui/popup.h
	src/gui/scrollarea.cpp
	src/gui/skin.cpp
	src/gui/skin.h
	src/gui/speechbubble.cpp
	src/gui/window.cpp
	src/gui/window.h
	src/localplayer.h
	src/main.cpp
	src/net/ea/playerhandler.cpp
	src/resources/ambientoverlay.h
	src/resources/dye.cpp
	src/resources/imagewriter.cpp
	src/resources/itemdb.cpp
	src/shopitem.cpp

+++ src/net/ea/playerhandler.cpp
@@ -211,9 +211,8 @@ void PlayerHandler::handleMessage(MessageIn &msg)
                 player_node->mX = x;
                 player_node->mY = y;
 
-                logger->log("Adjust scrolling by %d:%d",
-                        (int) scrollOffsetX,
-                        (int) scrollOffsetY);
+                logger->log("Adjust scrolling by %d:%d", (int) scrollOffsetX,
+                           (int) scrollOffsetY);
 
                 viewport->scrollBy(scrollOffsetX, scrollOffsetY);
             }
@@ -222,7 +221,7 @@ void PlayerHandler::handleMessage(MessageIn &msg)
         case SMSG_PLAYER_STAT_UPDATE_1:
             {
                 int type = msg.readInt16();
-                Uint32 value = msg.readInt32();
+                int value = msg.readInt32();
 
                 switch (type)
                 {
@@ -242,7 +241,7 @@ void PlayerHandler::handleMessage(MessageIn &msg)
                                  skillDialog->update();
                                  break;
                     case 0x0018:
-                                 if ((int) value >= player_node->getMaxWeight() / 2 &&
+                                 if (value >= player_node->getMaxWeight() / 2 &&
                                          player_node->getTotalWeight() <
                                          player_node->getMaxWeight() / 2)
                                  {
